fixed model

+++ src/main/java/com/example/demo/model/Assignment.java
@@ -2,7 +2,6 @@ package com.example.demo.model;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.time.LocalDate;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -33,7 +32,7 @@ public class Assignment {
   private String content;
 
   @Column(name = "passing_score")
-  private Integer passingScore;
+  private Float passingScore;
 
   @Column(name = "due_date")
   @DateTimeFormat(pattern = "yyyy-MM-dd")
@@ -51,7 +50,7 @@ public class Assignment {
   public Assignment() {
   }
 
-  public Assignment(Integer id, String name, String content, Integer passingScore, LocalDate dueDate, Material material) {
+  public Assignment(Integer id, String name, String content, Float passingScore, LocalDate dueDate, Material material) {
     this.id = id;
     this.name = name;
     this.content = content;
@@ -84,11 +83,11 @@ public class Assignment {
     this.content = content;
   }
 
-  public Integer getPassingScore() {
+  public Float getPassingScore() {
     return passingScore;
   }
 
-  public void setPassingScore(Integer passingScore) {
+  public void setPassingScore(Float passingScore) {
     this.passingScore = passingScore;
   }
